code to read local

+++ scripts/download_submission.py
@@ -5,7 +5,11 @@ from lxml import etree
 from conn import engine
 from tables import submissions
 from sqlalchemy import insert, update, select
+import os.path
 
+def open_or_download_file(url:str, accession_number:str):
+    if os.path.isfile(f'data/form4/{accession_number}.xml'):
+        return
 
 
 def save_submission(url: str, accession_number: str):
@@ -13,15 +17,19 @@ def save_submission(url: str, accession_number: str):
     xml_url = url.split("/")
     del xml_url[-2]
     xml_url = "/".join(xml_url)
-    response = requests.get(xml_url, headers=headers)
-
-    with open(f'data/form4/{accession_number}.xml', 'wb') as file:
-        raw_xml = response.content
-        parser = etree.XMLParser(remove_blank_text=True)
-        raw_xml = etree.XML(raw_xml, parser=parser)
-        file.write(etree.tostring(raw_xml))
-
-    return xmltodict.parse(response.content)
+    
+    if os.path.isfile(f'data/form4/{accession_number}.xml'):
+        with open(f'data/form4/{accession_number}.xml') as f:
+            raw_xml = f.read()
+    else:
+        response = requests.get(xml_url, headers=headers)
+        with open(f'data/form4/{accession_number}.xml', 'wb') as file:
+            raw_xml = response.content
+            parser = etree.XMLParser(remove_blank_text=True)
+            raw_xml = etree.XML(raw_xml, parser=parser)
+            raw_xml = etree.tostring(raw_xml)
+            file.write(raw_xml)
+    return xmltodict.parse(raw_xml)
 
 
 def update_submission(submission:dict, accession_number):
@@ -48,6 +56,7 @@ def get_submissions(limit:int):
     with engine.connect() as conn:
         query = select(submissions).where(submissions.c.owner_cik.is_(None)).order_by(submissions.c.ts.desc()).limit(limit)
         return conn.execute(query).all()
+    
 records = get_submissions(100)
 
 for record in records:
